updated to jogl 2.3.2

+++ src/maspack/render/GL/GL2/GL2Viewer.java
@@ -12,14 +12,14 @@ import java.awt.event.MouseWheelListener;
 import java.io.File;
 import java.util.LinkedList;
 
-import javax.media.opengl.GL;
-import javax.media.opengl.GL2;
-import javax.media.opengl.GL2GL3;
-import javax.media.opengl.GLAutoDrawable;
-import javax.media.opengl.GLCapabilities;
-import javax.media.opengl.GLContext;
-import javax.media.opengl.GLProfile;
-import javax.media.opengl.glu.GLU;
+import com.jogamp.opengl.GL;
+import com.jogamp.opengl.GL2;
+import com.jogamp.opengl.GL2GL3;
+import com.jogamp.opengl.GLAutoDrawable;
+import com.jogamp.opengl.GLCapabilities;
+import com.jogamp.opengl.GLContext;
+import com.jogamp.opengl.GLProfile;
+import com.jogamp.opengl.glu.GLU;
 import javax.swing.JFrame;
 import javax.swing.event.MouseInputListener;
 
@@ -2210,7 +2210,7 @@ public class GL2Viewer extends GLViewer implements HasProperties {
 
    private boolean setupHSVInterpolation (GL2 gl) {
       // create special HSV shader to interpolate colors in HSV space
-      int prog = GLHSVShader.getShaderProgram(gl);
+      long prog = GLHSVShader.getShaderProgram(gl);
       if (prog > 0) {
          gl.glUseProgramObjectARB (prog);
          return true;
