+++ src/maspack/render/GL/GL2/GLHSVShader.java
@@ -9,13 +9,13 @@ package maspack.render.GL.GL2;
 import java.nio.ByteBuffer;
 import java.nio.IntBuffer;
 
-import javax.media.opengl.GL2;
+import com.jogamp.opengl.GL2;
 
 // see http://www.lighthouse3d.com/opengl/glsl/index.php?intro
 
 public class GLHSVShader {
    private static boolean myInitialized = false;
-   private static int myProgram = -1;
+   private static long myProgram = -1;
 
    @SuppressWarnings("unused")
    private static String[] testProg = new String[] {
@@ -71,7 +71,7 @@ public class GLHSVShader {
       "}"
    };
 
-   public static void printInfoLog (GL2 gl, int shader) {
+   public static void printInfoLog (GL2 gl, long shader) {
       //int[] infologLength = new int[1];
       //int[] charsWritten = new int[1];
       IntBuffer ibuf = IntBuffer.allocate (100);
@@ -90,17 +90,17 @@ public class GLHSVShader {
       
    }
 
-   public static int getStatus (GL2 gl, int prog, int type) {
+   public static int getStatus (GL2 gl, long prog, int type) {
       IntBuffer ibuf = IntBuffer.allocate (100);
 
       gl.glGetObjectParameterivARB (prog, type, ibuf);
       return ibuf.get(0);      
    }
 
-   public static int createShaderProgram(GL2 gl) {
+   public static long createShaderProgram(GL2 gl) {
 
       System.out.println ("Initializing HSV shader ...");
-      int shader = gl.glCreateShaderObjectARB (GL2.GL_FRAGMENT_SHADER);
+      long shader = gl.glCreateShaderObjectARB (GL2.GL_FRAGMENT_SHADER);
 
       gl.glShaderSourceARB (shader, hsvProg.length, hsvProg, null);
       gl.glCompileShaderARB (shader);
@@ -110,7 +110,7 @@ public class GLHSVShader {
          printInfoLog (gl, shader);
          return -1;
       }
-      int prog = gl.glCreateProgramObjectARB();
+      long prog = gl.glCreateProgramObjectARB();
       gl.glAttachObjectARB(prog,shader);
       gl.glLinkProgramARB (prog);
       status = getStatus (gl, prog, GL2.GL_OBJECT_LINK_STATUS_ARB);
@@ -123,7 +123,7 @@ public class GLHSVShader {
       return prog;
    }
 
-   public static int getShaderProgram (GL2 gl) {
+   public static long getShaderProgram (GL2 gl) {
       if (!myInitialized) {
          myProgram = createShaderProgram (gl);
          myInitialized = true;
